Factor out new components, and implement basic positioning logic that accounts for hitting the top and right screen boundaries
