+++ browser/src/UI/components/QuickInfo.tsx
@@ -5,6 +5,8 @@ import { connect } from "react-redux"
 
 import { IState } from "./../State"
 
+import { CursorPositioner } from "./CursorPositioner"
+
 require("./QuickInfo.less") // tslint:disable-line no-var-requires
 
 export interface IQuickInfoProps {
@@ -14,141 +16,6 @@ export interface IQuickInfoProps {
     backgroundColor: string
     foregroundColor: string
 }
-
-const getOpenPosition = (state: IState): { backgroundColor: string, x: number, y: number, openFromTop: boolean } => {
-    const openFromTopPosition = state.cursorPixelY + (state.fontPixelHeight * 2)
-    const openFromBottomPosition = state.cursorPixelY - state.fontPixelHeight
-
-    const openFromTop = state.cursorPixelY < 75
-
-    const xPos = state.cursorPixelX - (state.fontPixelWidth / 2) - 2
-    const yPos = openFromTop ? openFromTopPosition : openFromBottomPosition
-
-    return {
-        x: xPos,
-        y: yPos,
-        openFromTop,
-        backgroundColor: state.backgroundColor,
-    }
-}
-
-export interface ICursorPositionerViewProps {
-    x: number
-    y: number
-
-    openFromTop: boolean
-
-    backgroundColor: string
-}
-
-export interface ICursorPositionerViewState {
-    measuredRect: ClientRect | null
-}
-
-/**
- * Helper component to position an element relative to the current cursor position
- */
-export class CursorPositionerView extends React.PureComponent<ICursorPositionerViewProps, ICursorPositionerViewState> {
-
-    constructor(props: ICursorPositionerViewProps) {
-        super(props)
-
-        this.state = {
-            measuredRect: null
-        }
-    }
-
-    private _measureElement(element: HTMLElement): void {
-        if (element) {
-            const rect = element.getBoundingClientRect()
-
-            if (!this.state.measuredRect
-                || this.state.measuredRect.width !== rect.width
-                || this.state.measuredRect.height !== rect.height) {
-
-                this.setState({
-                    measuredRect: rect
-                })
-            }
-        }
-    }
-
-    public render(): JSX.Element {
-        const containerStyle: React.CSSProperties = {
-            position: "absolute",
-            top: this.props.y.toString() + "px",
-            left: this.props.x.toString() + "px",
-        }
-
-        const openFromBottomStyle: React.CSSProperties = {
-            position: "absolute",
-            bottom: "0px",
-        }
-
-        const openFromTopStyle: React.CSSProperties = {
-            position: "absolute",
-            top: "0px",
-        }
-
-        const innerStyle = this.props.openFromTop ? openFromTopStyle : openFromBottomStyle
-
-        return <div style={containerStyle}>
-                <div style={innerStyle}>
-                    <div ref={(elem) => this._measureElement(elem) }>
-                        // <div>{!!this.state.measuredRect ? `${this.state.measuredRect.width}x${this.state.measuredRect.height}`: ""}</div>
-                        {this.props.children}
-                    </div>
-                 </div>
-                 <div style={{position: "absolute", top: "0px", left: "2px"}}>
-                     <Arrow direction={this.props.openFromTop ? ArrowDirection.Up : ArrowDirection.Down} size={10} color={this.props.backgroundColor} />
-                 </div>
-                </div>
-    }
-}
-
-export enum ArrowDirection {
-    Up = 0,
-    Down,
-}
-
-
-export interface IArrowProps {
-    size: number
-    color: string
-    direction: ArrowDirection
-}
-
-export const Arrow = (props: IArrowProps): JSX.Element => {
-
-    const transparentBorder = `${props.size * 0.8}px solid transparent`
-    const solidBorder = `${props.size}px solid ${props.color}`
-
-    const upArrowStyle = {
-        width: "0px",
-        height: "0px",
-        borderLeft: transparentBorder,
-        borderRight: transparentBorder,
-        borderBottom: solidBorder,
-    }
-
-    const downArrowStyle = {
-        width: "0px",
-        height: "0px",
-        borderLeft: transparentBorder,
-        borderRight: transparentBorder,
-        borderTop: solidBorder
-    }
-
-    const style = props.direction === ArrowDirection.Up ? upArrowStyle : downArrowStyle
-
-    return <div style={style}></div>
-}
-
-const mapStateToProps2 = (state: IState): ICursorPositionerViewProps => {
-    return getOpenPosition(state)
-}
-
-export const CursorPositioner = connect(mapStateToProps2)(CursorPositionerView)
 
 export class QuickInfo extends React.PureComponent<IQuickInfoProps, void> {
